Jikka.CPlusPlus.Language.Expr

Description

Expr module has the basic data types for C++ language. The data types are intended to use for the code generation.

Documentation

data NameHint

Constructors

LocalNameHint

LocalArgumentNameHint

LoopCounterNameHint

ConstantNameHint

FunctionNameHint

ArgumentNameHint

AdHocNameHint String

data VarName

Constructors

VarName OccName NameFlavour (Maybe NameHint)

formatVarName :: VarName -> String

newtype FunName

Constructors

FunName String

formatFunName :: FunName -> String

data Type

Constructors

TyAuto
auto

TyVoid
void

TyBool
bool

TyInt
int

TyInt32
int32_t

TyInt64
int64_t

TyTuple [Type]
std::tuple<T1, T2, ...>

TyVector Type
std::vector<T>

TyArray Type Integer
std::arrya<T, n>

TyString
std::string

TyFunction Type [Type]
std::function<Tr (T1, T2, ...)>

TyConvexHullTrick
jikka::convex_hull_trick

TySegmentTree Monoid'
atcoder::segtree<T, op, e>

TyIntValue Integer
an integer n for template parameters

data Monoid'

Constructors

MonoidIntPlus
\((\mathbb{Z}, +, 0)\)

MonoidIntMin
\((\mathrm{int64\_t}, \min, \mathrm{INT64\_MAX})\)

MonoidIntMax
\((\mathrm{int64\_t}, \max, \mathrm{INT64\_MIN})\)

MonoidIntGcd
\((\mathbb{Z}, \gcd, 0)\)

MonoidIntLcm
\((\mathbb{Z}, \mathrm{lcm}, 1)\)

data Literal

Constructors

LitInt32 Integer

LitInt64 Integer

LitBool Bool

LitChar Char

LitString String

data Function

Constructors

Function FunName [Type]
other functions

Method FunName
other methods

At
subscription e1[e2]

SetAt Type
updated array auto tmp = e1; tmp[e2] = e3; return tmp;

Cast Type
cast (T)e

StdTuple [Type]
functio std::tuple<T1, T2, ...>(e1, e2, ...)

StdGet Integer
function std::get<T, n>(e)

ArrayExt Type
std::array<T, n>{e1, e2, ..., en}

VecExt Type
std::vector<T>{e1, e2, ...}

VecCtor Type
constructors std::vector<T>() std::vector<T>(n) std::vector<T>(n, e)

Range
function std::vector<int> jikka::range(int n), which is similar to Python's range or Boost's boost::range

MethodSize
size method of std::vector<T>

ConvexHullTrickCtor
the constructor of jikka::convex_hull_trick

ConvexHullTrickCopyAddLine
This makes a copy of jikka::convex_hull_trick and updates it. This is removed at run.

SegmentTreeCtor Monoid'
the constructors of atcoder::segtree<T, op, e>

SegmentTreeCopySetPoint Monoid'
This makes a copy of atcoder::segtree<T, op, e> and updates it. This is removed at run.

data UnaryOp

Constructors

IntNop
+

Negate
-

BitNot
~

Not
! / not

Deref
*

data BinaryOp

Constructors

Add
+

Sub
-

Mul
*

Div
/

Mod
%

BitAnd
&

BitOr
|

BitXor
^

BitLeftShift
<<

BitRightShift
>>

And
&& / and

Or
|| / or

LessThan
<

LessEqual
<=

GreaterThan
>

GreaterEqual
>=

Equal
==

NotEqual
!=

data AssignOp

Constructors

SimpleAssign
=

AddAssign
+=

SubAssign
-=

MulAssign
*=

DivAssign
/=

ModAssign
%=

BitLeftShiftAssign
<<=

BitRightShiftAssign
>>=

BitAndAssign
&=

BitOrAssign
|=

BitXorAssign
^=

data Expr

Constructors

Var VarName

Lit Literal

UnOp UnaryOp Expr

BinOp BinaryOp Expr Expr

Cond Expr Expr Expr
e1 ? e2 : e3

Lam [(Type, VarName)] Type [Statement]
lambda expression [=](T1 x1, T2 x2, ...) -> Tr { stmt1; stmt2; ... }

Call Expr [Expr]
f(e1, e2, ...) for a callable f

Callable Function

data LeftExpr

Constructors

LeftVar VarName
x

LeftAt LeftExpr Expr
e[i]

LeftGet Integer LeftExpr
std::get<n>

data AssignExpr

Constructors

AssignExpr AssignOp LeftExpr Expr
e1 = e2

AssignIncr LeftExpr
++ e

AssignDecr LeftExpr
-- e

data DeclareRight

Constructors

DeclareDefault
T x;

DeclareCopy Expr
T x = e;

DeclareInitialize [Expr]
T x(e1, e2, ...);. This is only for better formatting. This should not be used while optimization phases.

data Statement

Constructors

ExprStatement Expr
e;

Block [Statement]
{ stmt1; stmts2; ...; }

If Expr [Statement] (Maybe [Statement])
if (e) { stmt1; stmts2; ...; } / if (e) { stmt1; stmts2; ...; } else { stmt1'; stmt2'; ...; }

For Type VarName Expr Expr AssignExpr [Statement]
for (T x = e1; e2; e3) { stmt1; stmts2; ...; }

ForEach Type VarName Expr [Statement]
for (T x : e) { stmt1; stmts2; ...; }

While Expr [Statement]
while (e) { stmt1; stmts2; ...; }

Declare Type VarName DeclareRight
Declarations with/witout initializations. See DeclareRight.

DeclareDestructure [VarName] Expr
auto [x1, x2, ...] = e;

Assign AssignExpr
e1 op= e2;

Assert Expr
assert (e);

Return Expr
return e;

data ToplevelStatement

Constructors

VarDef Type VarName Expr
const T x = e;

FunDef Type FunName [(Type, VarName)] [Statement]
T f(T1 x1, T2 x2, ...) { stmt1; stmt2; ... }

StaticAssert Expr String
static_assert(e, msg);

newtype Program

Constructors

Program

Fields

decls :: [ToplevelStatement]