Jikka.Common.Error

Description

Documentation

module Control.Monad.Except

error data types

data Responsibility

Constructors

UserMistake

ImplementationBug

data ErrorGroup

Constructors

LexicalError
It's impossible to split the given source text into tokens.

SyntaxError
It's impossible to construct AST from tokens.

SymbolError
There are undefined variables or functions in AST.

TypeError
It's impossible reconstruct types for AST.

SemanticError
other semantic erros

EvaluationError
User's program are not ready to evaluate.

RuntimeError
User's program failed while running.

AssertionError
User's program violates its assertion.

CommandLineError
The given command line arguments are not acceptable.

WrongInputError
User's program was correctly running but wrong input text is given.

InternalError
It's an bug of implementation.

data Error

Constructors

Error String

ErrorAppend Error Error

WithGroup ErrorGroup Error

WithWrapped String Error

WithLocation Loc Error

WithResponsibility Responsibility Error

general utilities for Except

wrapError :: MonadError e m => (e -> e) -> m a -> m a

wrapError' :: MonadError Error m => String -> m a -> m a

wrapAt :: MonadError Error m => Loc -> m a -> m a

wrapAt' :: MonadError Error m => Maybe Loc -> m a -> m a

maybeToError :: MonadError a m => a -> Maybe b -> m b

eitherToError :: MonadError a m => Either a b -> m b

utilities to report multiple errors

catchError' :: MonadError e m => m a -> m (Either e a)

catchError` is the inverse of liftError.

reportErrors :: MonadError Error m => [Either Error a] -> m [a]

reportErrors2 :: MonadError Error m => Either Error a -> Either Error b -> m (a, b)

reportErrors3 :: MonadError Error m => Either Error a -> Either Error b -> Either Error c -> m (a, b, c)

reportErrors4 :: MonadError Error m => Either Error a -> Either Error b -> Either Error c -> Either Error d -> m (a, b, c, d)

reportErrors5 :: MonadError Error m => Either Error a -> Either Error b -> Either Error c -> Either Error d -> Either Error e -> m (a, b, c, d, e)

function to construct errors

lexicalError :: String -> Error

lexicalErrorAt :: Loc -> String -> Error

syntaxError :: String -> Error

syntaxErrorAt :: Loc -> String -> Error

symbolError :: String -> Error

symbolErrorAt :: Loc -> String -> Error

typeError :: String -> Error

semanticError :: String -> Error

evaluationError :: String -> Error

runtimeError :: String -> Error

assertionError :: String -> Error

commandLineError :: String -> Error

wrongInputError :: String -> Error

internalError :: String -> Error

actions to throw errors

throwLexicalError :: MonadError Error m => String -> m a

throwLexicalErrorAt :: MonadError Error m => Loc -> String -> m a

throwSyntaxError :: MonadError Error m => String -> m a

throwSyntaxErrorAt :: MonadError Error m => Loc -> String -> m a

throwSyntaxErrorAt' :: MonadError Error m => Maybe Loc -> String -> m a

throwSymbolError :: MonadError Error m => String -> m a

throwSymbolErrorAt :: MonadError Error m => Loc -> String -> m a

throwSymbolErrorAt' :: MonadError Error m => Maybe Loc -> String -> m a

throwTypeError :: MonadError Error m => String -> m a

throwTypeErrorAt :: MonadError Error m => Loc -> String -> m a

throwTypeErrorAt' :: MonadError Error m => Maybe Loc -> String -> m a

throwSemanticError :: MonadError Error m => String -> m a

throwSemanticErrorAt :: MonadError Error m => Loc -> String -> m a

throwSemanticErrorAt' :: MonadError Error m => Maybe Loc -> String -> m a

throwEvaluationError :: MonadError Error m => String -> m a

throwRuntimeError :: MonadError Error m => String -> m a

throwRuntimeErrorAt :: MonadError Error m => Loc -> String -> m a

throwRuntimeErrorAt' :: MonadError Error m => Maybe Loc -> String -> m a

throwAssertionError :: MonadError Error m => String -> m a

throwCommandLineError :: MonadError Error m => String -> m a

throwWrongInputError :: MonadError Error m => String -> m a

throwInternalError :: MonadError Error m => String -> m a

throwInternalErrorAt :: MonadError Error m => Loc -> String -> m a

throwInternalErrorAt' :: MonadError Error m => Maybe Loc -> String -> m a

utilities for other types of errors

bug :: String -> a

todo :: String -> a