Jikka.Common.Format.Color

Documentation

data ColorFlag

Constructors

EnableColor

DisableColor

withColor :: ColorFlag -> Color -> String -> String

withBold :: ColorFlag -> String -> String

hGetColorFlag :: Handle -> IO ColorFlag

data Color

ANSI's eight standard colors. They come in two intensities, which are controlled by ColorIntensity. Many terminals allow the colors of the standard palette to be customised, so that, for example, setSGR [ SetColor Foreground Vivid Green ] may not result in bright green characters.

Constructors

Black

Red

Green

Yellow

Blue

Magenta

Cyan

White