Jikka.Common.Matrix

Documentation

data Matrix a

Matrix is data for matrices. It is guaranteed that internal arrays are not jagged arrays.

unMatrix :: Matrix a -> Vector (Vector a)

makeMatrix :: Vector (Vector a) -> Maybe (Matrix a)

makeMatrix' :: Vector (Vector a) -> Matrix a

matsize :: Matrix a -> (Int, Int)

matsize computes the size of a matrix.

matsize' :: Vector (Vector a) -> (Int, Int)

matsize` computes the size of a matrix. This assumes inputs are matrices (matcheck).

matcheck :: Vector (Vector a) -> Bool

matcheck checks a given vector of vectors is a matrix. That is, this returns False for jagged arrays.

matzero :: Num a => Int -> Matrix a

matone :: Num a => Int -> Matrix a

matadd :: Num a => Matrix a -> Matrix a -> Matrix a

matadd calculates the addition \(A + B\) of two matrices \(A, B\). This assumes sizes of inputs match.

matmul :: Num a => Matrix a -> Matrix a -> Matrix a

matmul calculates the multiplication \(A B\)of two matrices \(A, B\). This assumes sizes of inputs match.

matap :: Num a => Matrix a -> Vector a -> Vector a

matap calculates the multiplication \(A x\) of a matrix \(A\) and a vector \(x\). This assumes sizes of inputs match.

matscalar :: Num a => a -> Matrix a -> Matrix a

matpow :: (Show a, Num a) => Matrix a -> Integer -> Matrix a

matpow calculates the power \(A^k\) of a matrix \(A\) and a natural number \(k\). This assumes inputs are square matrices. This fails for \(k \lt 0\).