Jikka.Core.Convert.StrengthReduction

Description

StrengthReduction replaces strong functions in exprs with weaker functions. For example, this replace abs x with max x (- x).

Documentation

run :: MonadError Error m => Program -> m Program