Jikka.Core.Language.ArithmeticalExpr

Basic functions

data ArithmeticalExpr

parseArithmeticalExpr :: Expr -> ArithmeticalExpr

parseArithmeticalExpr converts a given expr to a normal form \(\sum_i \prod_j e _ {i,j})\). This assumes given exprs have the type \(\mathbf{int}\).

formatArithmeticalExpr :: ArithmeticalExpr -> Expr

integerArithmeticalExpr :: Integer -> ArithmeticalExpr

negateArithmeticalExpr :: ArithmeticalExpr -> ArithmeticalExpr

plusArithmeticalExpr :: ArithmeticalExpr -> ArithmeticalExpr -> ArithmeticalExpr

minusArithmeticalExpr :: ArithmeticalExpr -> ArithmeticalExpr -> ArithmeticalExpr

multArithmeticalExpr :: ArithmeticalExpr -> ArithmeticalExpr -> ArithmeticalExpr

isZeroArithmeticalExpr :: ArithmeticalExpr -> Bool

isOneArithmeticalExpr :: ArithmeticalExpr -> Bool

Advanced functions

unNPlusKPattern :: ArithmeticalExpr -> Maybe (VarName, Integer)

unNPlusKPattern recognizes a pattern of \(x + k\) for a variable \(x\) and an integer constant \(k \in \mathbb{Z}\).

makeVectorFromArithmeticalExpr :: Vector VarName -> ArithmeticalExpr -> Maybe (Vector ArithmeticalExpr, ArithmeticalExpr)

makeVectorFromArithmeticalExpr makes a vector \(f\) and a expr \(c\) from a given vector of variables \(x_0, x_1, \dots, x _ {n - 1}\) and a given expr \(e\) s.t. \(f\) and \(c\) don't have \(x_0, x_1, \dots, x _ {n - 1}\) as free variables and \(e = c + f \cdot (x_0, x_1, \dots, x _ {n - 1})\) holds. This assumes given variables and exprs have the type \(\mathbf{int}\).

The returned exprs are normalized with normalizeArithmeticalExpr.

makeAffineFunctionFromArithmeticalExpr :: VarName -> ArithmeticalExpr -> Maybe (ArithmeticalExpr, ArithmeticalExpr)

makeAffineFunctionFromArithmeticalExpr is a specialized version of makeVectorFromArithmeticalExpr. This function returns \(a, b\) for a given variable \(x\) and a given expr \(e = a x + b\) where \(a, b\) which doesn't use \(x\) free.

splitConstantFactorArithmeticalExpr :: ArithmeticalExpr -> (Integer, ArithmeticalExpr)

splitConstantFactorArithmeticalExpr finds \(k\) and \(e'\) for given \(e\) s.t. \(e = k e'\).