Jikka.Core.Language.FreeVars

Description

Documentation

isFreeVar :: VarName -> Expr -> Bool

isFreeVar checks if the given variable occurs in the tiven expr. This considers contexts.

>>> VarName "x" `isFreeVar` Lam (VarName "y") IntTy (Var (VarName "x"))
True

>>> VarName "x" `isFreeVar` Lam (VarName "x") IntTy (Var (VarName "x"))
False

isUnusedVar :: VarName -> Expr -> Bool

isUnusedVar is the negation of isFreeVar.

TODO: rename to isNonFreeVar?

isFreeVarOrScopedVar :: VarName -> Expr -> Bool

isFreeVarOrScopedVar checks if the given variable occurs in the tiven expr. This ignores contexts.

>>> VarName "x" `isFreeVarOrScopedVar` Lam (VarName "x") IntTy (Var (VarName "y"))
True

freeTyVars :: Type -> [TypeName]

findUnusedVarName :: VarName -> Expr -> VarName

findUnusedVarName' :: Expr -> VarName