Jikka.Core.Language.ModuloExpr

Basic functions

data ModuloExpr

parseModuloExpr :: Expr -> Maybe ModuloExpr

parseModuloExpr converts a given expr to a normal form \(\sum_i \prod_j e _ {i,j}) \bmod m\). This assumes given exprs have the type \(\mathbf{int}\).

formatTopModuloExpr :: ModuloExpr -> Expr

formatTopModuloExpr convert ModuloExpr to Expr with adding FloorMod at only the root.

formatBottomModuloExpr :: ModuloExpr -> Expr

formatBottomModuloExpr convert ModuloExpr to Expr with adding FloorMod at every nodes.

integerModuloExpr :: Modulo -> Integer -> ModuloExpr

negateModuloExpr :: ModuloExpr -> ModuloExpr

plusModuloExpr :: ModuloExpr -> ModuloExpr -> Maybe ModuloExpr

minusModuloExpr :: ModuloExpr -> ModuloExpr -> Maybe ModuloExpr

multModuloExpr :: ModuloExpr -> ModuloExpr -> Maybe ModuloExpr

isZeroModuloExpr :: ModuloExpr -> Bool

isOneModuloExpr :: ModuloExpr -> Bool

moduloOfModuloExpr :: ModuloExpr -> Expr

arithmeticExprFromModuloExpr :: ModuloExpr -> ArithmeticExpr

Utilities

newtype Modulo

Modulo is just a newtype to avoid mistakes that swapping left and right of mod-op.

Constructors

Modulo

Fields

unModulo :: Expr

formatModulo :: Modulo -> String

isModulo :: Expr -> Modulo -> Bool