Jikka.Core.Language.RewriteRules

Description

Documentation

newtype RewriteRule m

Constructors

RewriteRule ([(VarName, Type)] -> Expr -> m (Maybe Expr))

pureRewriteRule :: Monad m => ([(VarName, Type)] -> Expr -> Maybe Expr) -> RewriteRule m

simpleRewriteRule :: Monad m => (Expr -> Maybe Expr) -> RewriteRule m

applyRewriteRule :: MonadError Error m => RewriteRule m -> [(VarName, Type)] -> Expr -> StateT Integer m (Maybe Expr)

applyRewriteRule applies a given rule to a given expr. This rewrites on all sub-exprs of the given expr, and repeats to rewrite while it is possible.

This function is idempotent.

This function doesn't terminate when a given rewrite rule doesn't terminate.

applyRewriteRuleToplevelExpr :: MonadError Error m => RewriteRule m -> [(VarName, Type)] -> ToplevelExpr -> StateT Integer m (Maybe ToplevelExpr)

applyRewriteRuleProgram :: MonadError Error m => RewriteRule m -> Program -> m (Maybe Program)

applyRewriteRuleProgram' :: MonadError Error m => RewriteRule m -> Program -> m Program

traceRewriteRule :: Monad m => RewriteRule m -> RewriteRule m