Jikka.Core.Language.Vars

Description

Documentation

isFreeVar :: VarName -> Expr -> Bool

isFreeVar checks if the given variable occurs in the tiven expr. This considers contexts.

>>> VarName "x" `isFreeVar` Lam1 (VarName "y") IntTy (Var (VarName "x"))
True

>>> VarName "x" `isFreeVar` Lam1 (VarName "x") IntTy (Var (VarName "x"))
False

isUnusedVar :: VarName -> Expr -> Bool

isUnusedVar is the negation of isFreeVar.

TODO: rename to isNonFreeVar?

isFreeVarOrScopedVar :: VarName -> Expr -> Bool

isFreeVarOrScopedVar checks if the given variable occurs in the tiven expr. This ignores contexts.

>>> VarName "x" `isFreeVarOrScopedVar` Lam1 (VarName "x") IntTy (Var (VarName "y"))
True

findFreshVar :: Expr -> VarName

findFreshVar generates a variable which isn't free in the given expr. The return variable name may conflict if you ignores the context.

findFreshVar' :: [Expr] -> VarName

TODO: make a function to list free/scoped vars and use it to optimize the time complexity.

getAnonymousVar :: Int -> VarName

freeTyVars :: Type -> [TypeName]