Jikka.Language.Python.Typed.TypeInfer

Documentation

data TypeEnv

Constructors

TypeEnv

Fields

funenv :: [(FunName, ([Type], Type))]

varenv :: [(VarName, Type)]

emptyTypeEnv :: TypeEnv

infer' :: TypeEnv -> Expr -> Either String Type

infer :: TypeEnv -> Expr -> Type

infer extracts the type of the given expr, under the assumption that the expr is correctly typed.

check' :: TypeEnv -> Expr -> Either String Type

check :: TypeEnv -> Expr -> Either String ChurchType

check calculates the type of the given expr, with strict checking about their Church-style types. This function says nothing about Curry-style types.