Jikka.Python.Language.Expr

Description

Expr module has the basic data types for the standard Python. See the Python's ast module (https://docs.python.org/ja/3/library/ast.html#abstract-grammar) for reference.

Documentation

newtype Ident

Constructors

Ident String

unIdent :: Ident -> String

type Ident' = WithLoc Ident

data Constant

Constructors

ConstNone

ConstInt Integer

ConstBool Bool

ConstString String

ConstBytes [Int8]

ConstFloat Double

ConstImaginary Double

data Statement

Constructors

FunctionDef Ident' Arguments [Statement'] [Decorator] (Maybe Type')

AsyncFunctionDef Ident' Arguments [Statement'] [Decorator] (Maybe Type')

ClassDef Ident' [Expr'] [Keyword'] [Statement'] [Decorator]

Return (Maybe Expr')

Delete [Target']

Assign [Target'] Expr'

AugAssign Target' Operator Expr'

AnnAssign Target' Type' (Maybe Expr')

For Target' Expr' [Statement'] [Statement']
For target iter body orelse

AsyncFor Target' Expr' [Statement'] [Statement']

While Expr' [Statement'] [Statement']
While test body orelse

If Expr' [Statement'] [Statement']

With [WithItem] [Statement']

AsyncWith [WithItem] [Statement']

Raise (Maybe Expr') (Maybe Expr')
Raise exc cause represents raise exc from cause.

Try [Statement'] [ExceptHandler'] [Statement'] [Statement']
Try body handlers orelse finalbody

Assert Expr' (Maybe Expr')
Assert test msg

Import [Alias]

ImportFrom [Ident'] [Alias]

Global [Ident']

Nonlocal [Ident']

Expr' Expr'

Pass

Break

Continue

type Statement' = WithLoc Statement

data Expr

Constructors

BoolOp Expr' BoolOp Expr'

NamedExpr Target' Expr'
produced by the walrus operator :=

BinOp Expr' Operator Expr'

UnaryOp UnaryOp Expr'

Lambda Arguments Expr'

IfExp Expr' Expr' Expr'
IfExp test body orelse

Dict [(Maybe Expr', Expr')]
NULL key is for **d.

Set [Expr']

ListComp Expr' [Comprehension]

SetComp Expr' [Comprehension]

DictComp Expr' Expr' [Comprehension]

GeneratorExp Expr' [Comprehension]

Await Expr'

Yield (Maybe Expr')

YieldFrom Expr'

Compare Expr' [(CmpOp, Expr')]

Call Expr' [Expr'] [Keyword']

FormattedValue Expr' (Maybe Char) (Maybe Expr')
FormattedValue value conversion format_spec for f-strings

JoinedStr [Expr']

Constant Constant

Attribute Expr' Ident'
can appear in assignment context

Subscript Expr' Expr'
can appear in assignment context

Starred Expr'
can appear in assignment context

Name Ident'
can appear in assignment context

List [Expr']
can appear in assignment context

Tuple [Expr']
can appear in assignment context

Slice (Maybe Expr') (Maybe Expr') (Maybe Expr')
Slice lower upper step can appear only in Subscript

type Expr' = WithLoc Expr

type Target = Expr'

type Target' = Expr'

type Type = Expr'

type Type' = Expr'

type Decorator = Expr'

type Decorator' = Expr'

data ExprContext

Constructors

Load

Store

Del

data BoolOp

Constructors

And

Or

Implies
our extension

data Operator

Constructors

Add

Sub

Mult

MatMult

Div

FloorDiv

FloorMod

CeilDiv
our extension

CeilMod
our extension

Pow

BitLShift

BitRShift

BitOr

BitXor

BitAnd

Max
our extension

Min
our extension

data UnaryOp

Constructors

Invert
on int

Not
on bool

UAdd

USub

data CmpOp

Constructors

Eq'

NotEq

Lt

LtE

Gt

GtE

Is

IsNot

In

NotIn

data Comprehension

Comprehension target iter ifs is_async

Constructors

Comprehension

Fields

compTarget :: Target'

compIter :: Expr'

compIfs :: Maybe Expr'

data ExceptHandler

Constructors

ExceptHandler

Fields

exchType :: Maybe Type'

exchName :: Maybe Ident'

exchBody :: [Statement']

type ExceptHandler' = WithLoc ExceptHandler

data Arguments

Constructors

Arguments

Fields

argsPosonlyargs :: [Arg]

argsArgs :: [Arg]

argsVarargs :: Maybe Arg

argsKwonlyargs :: [Arg]

argsKwDefaults :: [Expr']

argsKwarg :: Maybe Arg

argsDefaults :: [Expr']

emptyArguments :: Arguments

type Arg = (Ident', Maybe Type')

type Keyword = (Maybe Ident', Expr')

NULL identifier is for **kwargs.

type Keyword' = WithLoc Keyword

type Alias = (Ident', Maybe Ident')

(name, asname). Alias is used for Import and ImportFrom.

type WithItem = (Expr', Maybe Expr')

(context_expr, optional_vars)

type Program = [Statement']