Jikka.RestrictedPython.Convert.TypeInfer

Description

Documentation

run :: (MonadAlpha m, MonadError Error m) => Program -> m Program

run infers types of given programs.

As the interface, you can understand this function does the following:

Finds a type environment \(\Gamma\) s.t. for all statement \(\mathrm{stmt}\) in the given program, \(\Gamma \vdash \mathrm{stmt}\) holds, and

Annotates each variable in the program using the \(\Gamma\).

In its implementation, this is just something like a Hindley-Milner type inference.

Requirements

There must be no name conflicts in given programs. They must be alpha-converted. (Alpha)

All names must be resolved. (ResolveBuiltin)

internal types and functions

data Equation

Constructors

TypeEquation Type Type (Maybe Loc)

TypeAssertion VarName' Type

formularizeProgram :: MonadAlpha m => Program -> m [Equation]

sortEquations :: [Equation] -> ([(Type, Type, Maybe Loc)], [(VarName', Type)])

mergeAssertions :: [(VarName', Type)] -> [(Type, Type, Maybe Loc)]

newtype Subst

Subst is type substituion. It's a mapping from type variables to their actual types.

Constructors

Subst

Fields

unSubst :: Map TypeName Type

subst :: Subst -> Type -> Type

solveEquations :: MonadError Error m => [(Type, Type, Maybe Loc)] -> m Subst

mapTypeProgram :: (Type -> Type) -> Program -> Program