Jikka.RestrictedPython.Language.Builtin

Documentation

builtinNames :: Set VarName

standardBuiltinNames :: Set VarName

additionalBuiltinNames :: Set VarName

resolveUniqueBuiltin :: (MonadAlpha m, MonadError Error m) => VarName' -> m Expr'

resolveUniqueBuiltin makes a builtin function from a variable name. However, this doesn't anything for ambiguous builtin functions. For example, the builtin function "max" is kept as a variable because it may be \(\mathbf{list}(\alpha) \to \alpha\), \(\alpha \times \alpha \to \alpha\), etc. and this function cannot resolve it.

resolveBuiltin :: (MonadAlpha m, MonadError Error m) => VarName' -> Int -> m Expr'

formatBuiltin :: Builtin -> String

typeBuiltin :: Builtin -> Type