Jikka.RestrictedPython.Language.Util

generating symbols

genType :: MonadAlpha m => m Type

genVarName :: MonadAlpha m => VarName' -> m VarName'

genVarName' :: MonadAlpha m => m VarName'

free variables

freeTyVars :: Type -> [TypeName]

freeVars :: Expr' -> [VarName]

freeVars` reports all free variables.

freeVars' :: Expr' -> [VarName']

freeVars` reports all free variables with their locations, i.e. occurrences. For examples, x + x and x have the same free variables x but they have different sets of occurrences of free variable.

freeVarsTarget :: Target' -> [VarName]

freeVarsTarget' :: Target' -> [VarName']

return-statements

doesAlwaysReturn :: Statement -> Bool

doesPossiblyReturn :: Statement -> Bool

traversing statements

mapStatement :: (Statement -> [Statement]) -> Program -> Program

mapStatementM :: Monad m => (Statement -> m [Statement]) -> Program -> m Program

mapStatementM replaces all statements in a given program using a given function. This may breaks various constraints.

mapLargeStatement :: (Expr' -> [Statement] -> [Statement] -> [Statement]) -> (Target' -> Expr' -> [Statement] -> [Statement]) -> Program -> Program

mapLargeStatementM :: Monad m => (Expr' -> [Statement] -> [Statement] -> m [Statement]) -> (Target' -> Expr' -> [Statement] -> m [Statement]) -> Program -> m Program

mapStatements :: ([Statement] -> [Statement]) -> Program -> Program

mapStatementsM :: Monad m => ([Statement] -> m [Statement]) -> Program -> m Program

listStatements :: Program -> [Statement]

traversing sub exprs

mapSubExprM :: Monad m => (Expr' -> m Expr') -> Expr' -> m Expr'

mapSubExprM replaces all exprs in a given expr using a given function. This may breaks various constraints.

mapSubExpr :: (Expr' -> Expr') -> Expr' -> Expr'

listSubExprs :: Expr' -> [Expr']

traversing exprs

mapExprTargetM :: Monad m => (Expr' -> m Expr') -> Target' -> m Target'

mapExprStatementM :: Monad m => (Expr' -> m Expr') -> Statement -> m Statement

mapExprM :: Monad m => (Expr' -> m Expr') -> Program -> m Program

listExprs :: Program -> [Expr']

exprs

hasFunctionCall :: Expr' -> Bool

isSmallExpr :: Expr' -> Bool

isSmallExpr is true if the evaluation of a given expr trivially terminates.

dropLocation :: Expr' -> Expr'

targets

targetVars :: Target' -> [VarName]

targetVars' :: Target' -> [VarName']

hasSubscriptTrg :: Target' -> Bool

hasBareNameTrg :: Target' -> Bool

exprToTarget :: Expr' -> Maybe Target'

targetToExpr :: Target' -> Expr'

programs

toplevelMainDef :: [Statement] -> Program