Jikka.RestrictedPython.Language.VariableAnalysis

Documentation

newtype ReadList

Constructors

ReadList [VarName]

newtype WriteList

Constructors

WriteList [VarName]

haveWriteReadIntersection :: WriteList -> ReadList -> Bool

analyzeExpr :: Expr' -> ReadList

analyzeTargetRead :: Target' -> ReadList

analyzeTargetWrite :: Target' -> WriteList

analyzeStatementGeneric :: Bool -> Statement -> (ReadList, WriteList)

analyzeStatementsGeneric :: Bool -> [Statement] -> (ReadList, WriteList)

analyzeStatementMax :: Statement -> (ReadList, WriteList)

analyzeStatementMax returns lists of variables which are possibly read or written in given statements.

analyzeStatementsMax :: [Statement] -> (ReadList, WriteList)

analyzeStatementMin :: Statement -> (ReadList, WriteList)

analyzeStatementMin returns lists of variables which are always read or written in given statements.

analyzeStatementsMin :: [Statement] -> (ReadList, WriteList)