Jikka.CPlusPlus.Language.Expr

Description

Expr module has the basic data types for C++ language. The data types are intended to use for the code generation.

Documentation

newtype VarName

Constructors

VarName

Fields

unVarName :: String

newtype FunName

Constructors

FunName

Fields

unFunName :: String

data Type

Constructors

TyAuto

TyVoid

TyBool

TyInt32

TyInt64

TyTuple [Type]

TyVector Type

TyArray Type Integer

data Literal

Constructors

LitInt32 Integer

LitInt64 Integer

LitBool Bool

data Function

Constructors

Callable Expr

Function FunName [Type]

Method Expr FunName

data UnaryOp

Constructors

IntNop

Negate

BitNot

Not

data BinaryOp

Constructors

Add

Sub

Mul

Div

Mod

BitAnd

BitOr

BitXor

BitLeftShift

BitRightShift

And

Or

LessThan

LessEqual

GreaterThan

GreaterEqual

Equal

NotEqual

data AssignOp

Constructors

SimpleAssign

AddAssign

SubAssign

MulAssign

DivAssign

ModAssign

BitLeftShiftAssign

BitRightShiftAssign

BitAndAssign

BitOrAssign

BitXorAssign

data Expr

Constructors

Var VarName

Lit Literal

UnOp UnaryOp Expr

BinOp BinaryOp Expr Expr

Cond Expr Expr Expr

Lam [(Type, VarName)] Type [Statement]

Call Function [Expr]

VecExt Type [Expr]

At Expr Expr

Cast Type Expr

data LeftExpr

Constructors

LeftVar VarName

LeftAt LeftExpr Expr

data AssignExpr

Constructors

AssignExpr AssignOp LeftExpr Expr

AssignIncr LeftExpr

AssignDecr LeftExpr

data Statement

Constructors

ExprStatement Expr

Block [Statement]

If Expr [Statement] (Maybe [Statement])

For Type VarName Expr Expr AssignExpr [Statement]

ForEach Type VarName Expr [Statement]

While Expr [Statement]

Declare Type VarName (Maybe Expr)

Assign AssignExpr

Assert Expr

Return Expr

data ToplevelStatement

Constructors

VarDef Type VarName Expr

FunDef Type VarName [(Type, VarName)] [Statement]

newtype Program

Constructors

Program

Fields

decls :: [ToplevelStatement]