Jikka.Core.Language.ArithmeticalExpr

Documentation

data ProductExpr

Constructors

ProductExpr

Fields

productExprConst :: Integer

productExprList :: [Expr]

data SumExpr

Constructors

SumExpr

Fields

sumExprList :: [ProductExpr]

sumExprConst :: Integer

type ArithmeticalExpr = SumExpr

oneProductExpr :: ProductExpr

negateProductExpr :: ProductExpr -> ProductExpr

multProductExpr :: ProductExpr -> ProductExpr -> ProductExpr

parseProductExpr :: Expr -> ProductExpr

sumExprFromProductExpr :: ProductExpr -> SumExpr

sumExprFromInteger :: Integer -> SumExpr

negateSumExpr :: SumExpr -> SumExpr

plusSumExpr :: SumExpr -> SumExpr -> SumExpr

multSumExpr :: SumExpr -> SumExpr -> SumExpr

parseSumExpr :: Expr -> SumExpr

parseArithmeticalExpr :: Expr -> ArithmeticalExpr

parseArithmeticalExpr converts a given expr to a normal form \(\sum_i \prod_j e _ {i,j})\). This assumes given exprs have the type \(\mathbf{int}\).

formatProductExpr :: ProductExpr -> Expr

formatSumExpr :: SumExpr -> Expr

formatArithmeticalExpr :: ArithmeticalExpr -> Expr

normalizeProductExpr :: ProductExpr -> ProductExpr

normalizeSumExpr :: SumExpr -> SumExpr

normalizeArithmeticalExpr :: ArithmeticalExpr -> ArithmeticalExpr

makeVectorFromArithmeticalExpr :: Vector VarName -> ArithmeticalExpr -> Maybe (Vector ArithmeticalExpr, ArithmeticalExpr)

makeVectorFromArithmeticalExpr makes a vector \(f\) and a expr \(c\) from a given vector of variables \(x_0, x_1, \dots, x _ {n - 1}\) and a given expr \(e\) s.t. \(f\) and \(c\) don't have \(x_0, x_1, \dots, x _ {n - 1}\) as free variables and \(e = c + f \cdot (x_0, x_1, \dots, x _ {n - 1})\) holds. This assumes given variables and exprs have the type \(\mathbf{int}\).

The returned exprs are normalized with normalizeArithmeticalExpr.

isZeroArithmeticalExpr :: ArithmeticalExpr -> Bool

isOneArithmeticalExpr :: ArithmeticalExpr -> Bool

unNPlusKPattern :: ArithmeticalExpr -> Maybe (VarName, Integer)

unNPlusKPattern recognizes a pattern of \(x + k\) for a variable \(x\) and an integer constant \(k \in \mathbb{Z}\).

makeAffineFunctionFromArithmeticalExpr :: VarName -> ArithmeticalExpr -> Maybe (ArithmeticalExpr, ArithmeticalExpr)

makeAffineFunctionFromArithmeticalExpr is a specialized version of makeVectorFromArithmeticalExpr. This function returns \(a, b\) for a given variable \(x\) and a given expr \(e = a x + b\) where \(a, b\) which doesn't use \(x\) free.