Jikka.RestrictedPython.Language.Lint

Documentation

makeEnsureProgram :: MonadError Error m => (Program -> Bool) -> String -> Program -> m ()

hasSubscriptionInLoopCounters :: Program -> Bool

hasSubscriptionInLoopCounters checks that there are SubscriptTrg in loop counters of for-loops. This includes loop counters of ListComp. For example, the followings has such subscriptions.

for a[0] in range(100):
    pass
return a[0]  # => 99

a = [0]
b = [0 for a[0] in range(100)]
return a[0]  # => 99

NOTE: This is allowd in the standard Python.

doesntHaveSubscriptionInLoopCounters :: Program -> Bool

ensureDoesntHaveSubscriptionInLoopCounters :: MonadError Error m => Program -> m ()

hasNameLeakOfLoopCounters :: Program -> Bool

hasNameLeakOfLoopCounters checks that there are leaks of loop counters of for-loops. For example, the following has a leak.

for i in range(100):
    pass
return i  # => 100

doesntHaveNameLeakOfLoopCounters :: Program -> Bool

ensureDoesntHaveNameLeakOfLoopCounters :: MonadError Error m => Program -> m ()

hasAssignmentToLoopCounters :: Program -> Bool

hasAssignmentToLoopCounters checks that there are assignments to loop counters of for-loops. For example, the following has the assignment.

for i in range(100):
    i += 1

doesntHaveAssignmentToLoopCounters :: Program -> Bool

ensureDoesntHaveAssignmentToLoopCounters :: MonadError Error m => Program -> m ()

hasAssignToLoopIterators :: Program -> Bool

hasAssignToLoopIterators checks that there are assignments to loop iterators of for-loops. For example, the followings have the assignments.

a = list(range(10))
for i in a:
    a[5] = i

a = 0
for i in f(a):
    a += i

doesntHaveAssignmentToLoopIterators :: Program -> Bool

ensureDoesntHaveAssignmentToLoopIterators :: MonadError Error m => Program -> m ()

hasReturnInLoops :: Program -> Bool

hasReturnInLoops checks that there are return-statements in for-loops. For example, the following has such a return-statement.

a = list(range(10))
for i in a:
    return True

doesntHaveReturnInLoops :: Program -> Bool

ensureDoesntHaveReturnInLoops :: MonadError Error m => Program -> m ()

hasMixedAssignment :: Program -> Bool

hasMixedAssignment checks that there are assignments which assign to both of bare variables and subscripted variables. For example, the following is such an assignment.

a, b[0] = list(range(10))

NOTE: this doesn't check loop counters of For or ListComp.

doesntHaveMixedAssignment :: Program -> Bool

ensureDoesntHaveMixedAssignment :: MonadError Error m => Program -> m ()